requirements.txt
absl-py
altair
astunparse
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
decorator
filelock
flatbuffers
fonttools
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
Jinja2
jsonschema
jsonschema-specifications
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
networkx
numpy
oauthlib
opencv-python-headless
opt-einsum
packaging
pandas
Pillow
protobuf
pyarrow
pyasn1
pyasn1-modules
pydeck
Pygments
Pympler
pyparsing
python-dateutil
pytz
pytz-deprecation-shim
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
six
smmap
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
toml
toolz
torch
torchvision
tornado
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
Werkzeug
wrapt
zipp